Move more config related code to config module
